Updated LibLoader, lib's hooks, and added in a draft of libAutoUpdate (unfinished)

+++ LibLoader.mod.gml
@@ -11,20 +11,23 @@ if(!mod_exists("mod", "lib")){
 	//(load this mod in the first txt file, then load a second txt file and have the allowmod command in that second txt file)
 	while(!mod_sideload()){wait 1;}
 
-	//downloading lib (no version checking because that would slow the process further)
-	//I delete for safety, there's a chance it isn't necessary
-	file_delete("../../mods/lib/lib.mod.gml");
-	file_delete("../../mods/lib/main.txt");
-	file_delete("../../mods/lib/main2.txt");
-	while (file_exists("../../mods/lib/lib.mod.gml")) {wait 1;}
-	while (file_exists("../../mods/lib/main.txt")) {wait 1;}
-	while (file_exists("../../mods/lib/main2.txt")) {wait 1;}
-	file_download(URL + "lib.mod.gml", "../../mods/lib/lib.mod.gml");
-	file_download(URL + "main.txt", "../../mods/lib/main.txt");
-	file_download(URL + "main2.txt", "../../mods/lib/main2.txt");
-	while (!file_loaded("../../mods/lib/lib.mod.gml")) {wait 1;}
-	while (!file_loaded("../../mods/lib/main.txt")) {wait 1;}
-	while (!file_loaded("../../mods/lib/main2.txt")) {wait 1;}
+	//Don't download anything if you're in multiplayer
+	if(!player_is_active(1) && !player_is_active(2) && !player_is_active(3)){
+		//downloading lib (no version checking because that would slow the process further)
+		//I delete for safety, there's a chance it isn't necessary
+		file_delete("../../mods/lib/lib.mod.gml");
+		file_delete("../../mods/lib/main.txt");
+		file_delete("../../mods/lib/main2.txt");
+		while (file_exists("../../mods/lib/lib.mod.gml")) {wait 1;}
+		while (file_exists("../../mods/lib/main.txt")) {wait 1;}
+		while (file_exists("../../mods/lib/main2.txt")) {wait 1;}
+		file_download(URL + "lib.mod.gml", "../../mods/lib/lib.mod.gml");
+		file_download(URL + "main.txt", "../../mods/lib/main.txt");
+		file_download(URL + "main2.txt", "../../mods/lib/main2.txt");
+		while (!file_loaded("../../mods/lib/lib.mod.gml")) {wait 1;}
+		while (!file_loaded("../../mods/lib/main.txt")) {wait 1;}
+		while (!file_loaded("../../mods/lib/main2.txt")) {wait 1;}
+	}
 
 	mod_loadtext("../../mods/lib/main.txt");
 }
